src/requirements.txt
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.12
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.0
urllib3==1.25.9
wrapt==1.12.1
